Add page to add babies

+++ src/Mde.Project.Mobile/Mde.Project.Mobile/Domain/Models/Baby.cs
@@ -8,8 +8,8 @@ namespace Mde.Project.Mobile.Domain.Models
     public class Baby : BaseModel
     {
         public string FirstName { get; set; }
-        public int Height { get; set; }
-        public int Weight { get; set; }
+        public double Height { get; set; }
+        public double Weight { get; set; }
         public Guid MotherId { get; set; }
         public Mother Mother { get; set; }
         public ICollection<Memory> Memories { get; set; }
