+++ src/Mde.Project.Mobile/Mde.Project.Mobile/Domain/Services/Interfaces/IBabyService.cs
@@ -9,7 +9,7 @@ namespace Mde.Project.Mobile.Domain.Services.Interfaces
     public interface IBabyService
     {
         Task<List<Baby>> GetBabies();
-        Task CreateBaby(string firstName, int height, int weight, Guid motherId);
+        Task CreateBaby(string firstName, double height, double weight, Guid motherId);
         Task UpdateBaby(Baby baby);
         Task DeleteBaby(Baby baby);
     }
